add variant and size composables, cleanup components
